feat: rename default tab settings[WTEL-9619](https://webitel.atlassian.net/browse/WTEL-9619)
