+++ src/enums/DefaultWorkspaceTabSettings/DefaultWorkspaceTabSettings.ts
